+++ apps/sos/src/syscall/mmap.cpp
@@ -16,7 +16,7 @@ async::future<int> brk(std::weak_ptr<process::Process> /*process*/, memory::vadd
     throw std::system_error(ENOSYS, std::system_category(), "brk() not implemented");
 }
 
-async::future<int> mmap2(std::weak_ptr<process::Process> process, memory::vaddr_t addr, size_t length, int prot, int flags, int /*fd*/, off_t /*offset*/) {
+async::future<int> mmap2(std::weak_ptr<process::Process> process, memory::vaddr_t addr, size_t length, int prot, int flags, int /*fd*/, int /*offset*/) {
     if (memory::pageAlign(addr) != addr || memory::pageAlign(length) != length)
         throw std::invalid_argument("Invalid page or length alignment");
